agent/requirements.txt
langchain==0.3.26
langgraph==0.4.10
langsmith==0.4.4
fastapi>=0.115.5,<1.0.0
uvicorn>=0.29.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
langgraph-cli[inmem]==0.3.3
google-generativeai>=0.8.3
